requirements.txt
# torch==1.8.1+cu111 torchvision==0.9.1+cu111 torchaudio==0.8.1 -f https://download.pytorch.org/whl/torch_stable.html
h5py==2.10.0
sklearn
tqdm
optuna
faiss-gpu
tensorboard
plotly
pandas
matplotlib
torchsummary
nuscenes-devkit